% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bridges.R
\name{st_bridges}
\alias{st_bridges}
\title{Create contiguityies when islands are present}
\usage{
st_bridges(
  df,
  geom_col_name,
  remove_islands = FALSE,
  link_islands_k = 1,
  nb_structure = "list",
  add_to_dataframe = TRUE
)
}
\arguments{
\item{add_to_dataframe}{}
}
\description{
Create contiguityies when islands are present
}
