% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_check_islands.R
\name{st_check_islands}
\alias{st_check_islands}
\title{Examine contiguity actions performed on islands}
\usage{
st_check_islands(data)
}
\arguments{
\item{data}{an sf dataframe with a neighbourhood column called "nb"}
}
\value{
a dataframe reporting non-contiguous connections made by st_bridges()
}
\description{
Examine contiguity actions performed on islands
}
\examples{
ukdata <- readRDS(system.file("extdata", "ukdata.rds", package="sfislands"))
st_bridges(ukdata,"constituency_name") |>
st_check_islands()
}
