% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_force_cut_nb.R
\name{st_force_cut_nb}
\alias{st_force_cut_nb}
\title{Remove contiguity between pairs of areas}
\usage{
st_force_cut_nb(nb, x = NULL, y = NULL, xy_df = NULL)
}
\arguments{
\item{nb}{a neighbourhood `"list"` or `"matrix"`, or an `sf` dataframe with a neighbourhood column called `"nb"`.}

\item{x}{name or number of first area (optional if `xy_df` is provided).}

\item{y}{name or number of second area (optional if `xy_df` is provided).}

\item{xy_df}{(optional) a dataframe with two columns: `x` and `y`, containing the names or numbers of areas to cut.}
}
\value{
An amended neighbourhood `"list"`, `"matrix"`, or `sf` dataframe with a neighbourhood column called `"nb"`.
}
\description{
Remove contiguity between pairs of areas
}
\examples{
# For individual x and y
st_force_cut_nb(uk_election, x = "Ynys Mon", y = "Arfon")

# For multiple x and y pairs
xy_df <- data.frame(x = c("Ynys Mon", "Gower"), y = c("Arfon", "Cardiff"))
st_force_cut_nb(uk_election, xy_df = xy_df)
}
