% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_manual_cut_nb.R
\name{st_manual_cut_nb}
\alias{st_manual_cut_nb}
\title{Manual remove contiguity between two areas}
\usage{
st_manual_cut_nb(nb, x, y)
}
\arguments{
\item{y}{}
}
\description{
Manual remove contiguity between two areas
}
