% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_manual_join_nb.R
\name{st_manual_join_nb}
\alias{st_manual_join_nb}
\title{Manually enforce contiguity between two areas}
\usage{
st_manual_join_nb(nb, x, y)
}
\arguments{
\item{y}{}
}
\description{
Manually enforce contiguity between two areas
}
