% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_quickmap_contigs.R
\name{st_quickmap_contigs}
\alias{st_quickmap_contigs}
\title{Visualise a contiguity structure on a map}
\usage{
st_quickmap_contigs(
  nbsf,
  linkcol = "dodgerblue",
  bordercol = "gray7",
  pointcol = "darkred",
  fillcol = "gray95",
  linksize = 0.2,
  bordersize = 0.1,
  pointsize = 0.8,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{subtitle}{}
}
\description{
Visualise a contiguity structure on a map
}
