% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_quickmap_preds.R
\name{st_quickmap_preds}
\alias{st_quickmap_preds}
\title{Visualilse the predictions of an `mgcv` model}
\usage{
st_quickmap_preds(output)
}
\arguments{
\item{output}{an augmented sf dataframe produced by st_augment()}
}
\value{
a list of ggplots
}
\description{
Visualilse the predictions of an `mgcv` model
}
\examples{
prepdata <- st_bridges(uk_election,"constituency_name")
mgcv::gam(health_not_good ~
  s(constituency_name, bs='mrf', xt=list(nb=prepdata$nb), k=100),data=prepdata, method="REML") |>
st_augment(uk_election) |>
st_quickmap_preds()
}
