% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_election.R
\docType{data}
\name{uk_election}
\alias{uk_election}
\title{UK election data}
\format{
## `uk_election`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{country}{Country name}
  \item{iso2, iso3}{2 & 3 letter ISO country codes}
  \item{year}{Year}
  ...
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
uk_election
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
