% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_augment.R
\name{st_augment}
\alias{st_augment}
\title{Augment dataframe with predictions of model}
\usage{
st_augment(model, df)
}
\arguments{
\item{model}{an `mgcv` model}

\item{df}{the sf data frame to be augmented with model preditions}
}
\value{
an augmented sd dataframe
}
\description{
Augment dataframe with predictions of model
}
