% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bridges.R
\name{st_bridges}
\alias{st_bridges}
\title{Create contiguities when islands are present}
\usage{
st_bridges(
  df,
  geom_col_name,
  remove_islands = FALSE,
  link_islands_k = 1,
  nb_structure = "list",
  add_to_dataframe = TRUE
)
}
\arguments{
\item{df}{an sf or sfc object}

\item{geom_col_name}{name of column containing names (or unique identifiers) for each row}

\item{remove_islands}{TRUE or FALSE (default)}

\item{link_islands_k}{an integer, k. Each island should be connected to the k nearest units}

\item{nb_structure}{either "list" (default) or "matrix"}

\item{add_to_dataframe}{TRUE (default) or FALSE. Whether or not to augment existing df with output as "nb" column}
}
\value{
a named neighbourhood list or matrix, optionally included as a column in an sf dataframe
}
\description{
Create contiguities when islands are present
}
\examples{
ukdata <- readRDS(system.file("extdata", "ukdata.rds", package="sfislands"))
st_bridges(ukdata,"constituency_name")
}
