% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_check_islands.R
\name{st_check_islands}
\alias{st_check_islands}
\title{Examine contiguity actions performed on islands}
\usage{
st_check_islands(data)
}
\arguments{
\item{data}{}
}
\description{
Examine contiguity actions performed on islands
}
