% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_manual_join_nb.R
\name{st_manual_join_nb}
\alias{st_manual_join_nb}
\title{Manually enforce contiguity between two areas}
\usage{
st_manual_join_nb(nb, x, y)
}
\arguments{
\item{nb}{a neighbourhood list or matrix, or an sf dataframe with a neighbourhood column called "nb"}

\item{x}{name or number of first area}

\item{y}{name or number of second area}
}
\value{
an amended neighbourhood list, matrix, or sf dataframe with a neighbourhood column called "nb"
}
\description{
Manually enforce contiguity between two areas
}
