% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_quickmap_nb.R
\name{st_quickmap_nb}
\alias{st_quickmap_nb}
\title{Visualise a neighbourhood structure on a map}
\usage{
st_quickmap_nb(
  nbsf,
  linkcol = "dodgerblue",
  bordercol = "gray7",
  pointcol = "darkred",
  fillcol = "gray95",
  linksize = 0.2,
  bordersize = 0.1,
  pointsize = 0.8,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{nbsf}{an sf dataframe with a neighbourhood column called "nb"}

\item{linkcol}{colour of lines connecting neighbours}

\item{bordercol}{colour of boundary lines between areas}

\item{pointcol}{colour of centroid points}

\item{fillcol}{fill of areas}

\item{linksize}{linewidth of lines connecting neighbours}

\item{bordersize}{linewidth of borders between areas}

\item{pointsize}{sie of centroid points}

\item{title}{plot title}

\item{subtitle}{plot subtitle}
}
\value{
ggplot of areas and neighbourhood structure
}
\description{
Visualise a neighbourhood structure on a map
}
\examples{
st_bridges(uk_election,"constituency_name") |>
st_quickmap_nb()
}
