% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_quickmap_preds.R
\name{st_quickmap_preds}
\alias{st_quickmap_preds}
\title{Visualilse the predictions of an `mgcv` model}
\usage{
st_quickmap_preds(output)
}
\arguments{
\item{output}{}
}
\description{
Visualilse the predictions of an `mgcv` model
}
