% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_quickmap_preds.R
\name{st_quickmap_preds}
\alias{st_quickmap_preds}
\title{Visualilse the predictions of an `mgcv` model}
\usage{
st_quickmap_preds(output)
}
\arguments{
\item{output}{an augmented sf dataframe produced by st_augment()}
}
\value{
a list of ggplots
}
\description{
Visualilse the predictions of an `mgcv` model
}
\examples{
ukdata <- readRDS(system.file("extdata", "ukdata.rds", package="sfislands")) |>
st_bridges("constituency_name")
mgcv::gam(health_not_good ~
  s(constituency_name, bs='mrf', xt=list(nb=ukdata$nb), k=100),data=ukdata, method="REML") |>
st_augment(ukdata) |>
st_quickmap_preds()
}
