% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_quickmap_preds.R
\name{st_quickmap_preds}
\alias{st_quickmap_preds}
\title{Visualilse the predictions of an `mgcv` model}
\usage{
st_quickmap_preds(
  output,
  scale_low = "firebrick4",
  scale_mid = "white",
  scale_high = "darkblue",
  scale_midpoint = 0,
  borderwidth = 0.05,
  bordercol = "black",
  legendlimits = "individual"
)
}
\arguments{
\item{output}{an augmented sf dataframe produced by st_augment()}

\item{scale_low}{fill of lowest extreme of scale}

\item{scale_mid}{fill of midpoint of scale}

\item{scale_high}{fill of highest extreme of scale}

\item{scale_midpoint}{value of midpoint of scale}

\item{borderwidth}{linewidth of borders between units}

\item{bordercol}{colour of borders between units}

\item{limits}{"individual" (default) each plot scaled to individual min-max. "minmax" means all plot have a common min-max according to the global min-max}
}
\value{
a list of ggplots
}
\description{
Visualilse the predictions of an `mgcv` model
}
\examples{
prepdata <- st_bridges(uk_election,"constituency_name")
mgcv::gam(health_not_good ~
  s(constituency_name, bs='mrf', xt=list(nb=prepdata$nb), k=100),data=prepdata, method="REML") |>
st_augment(uk_election) |>
st_quickmap_preds()
}
